+++ CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Java] Fix step match arguments highlighting for ambiguous steps ([#28](https://github.com/cucumber/pretty-formatter/pull/28))
 
 ### Added
-- Update dependency @cucumber/messages up to v29
-- Update dependency io.cucumber:messages up to v29
+- Update dependency @cucumber/query to v14
+- Update dependency @cucumber/messages to v29
+- Update dependency io.cucumber:query to v14
+- Update dependency io.cucumber:messages to v29
 
 ### Fixed
 - [JavaScript] Fix repeated step arguments ([#21](https://github.com/cucumber/pretty-formatter/pull/21))